services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
python-multipart==0.0.9
Pillow==10.3.0
cryptography==42.0.7
